Add DependentData: dependent features AND samples
The data generating process is essentially a Markov chain $p(x_1)
p(x_2|x_1) p(x_3|x_2) ...$. The conditionals are defined by matrices
that are sparse, generated by sampling binary coefficients and
filling these non-negative coefficients with a dirichlet. That's for
dependent features.

For dependent samples, in a very quick summary, we try to resample only
the last features (say, $p(x_4)$ if there are 4 features). When it
always yield the same vector, we go up a level and sample the feature
above ($p(x_3)$) until we reach $x_1$. There is a "noisy diagonal"
matrix that defines how to sample $p(x_1'|x_1)$.

To make these work, I needed to refactor the parsing a bit.