Logit squashing + adaptive entropy coefficient
- Logit squashing refers to putting a sigmoid on all the logits and
adding a small term so that the sender cannot really maxout the
probability all the way to 1. This is very crude and does not adapt to
vocabulary size. But the hope is that gradients never die.
- The entropy coefficient now looks e^(a*loss.detach()+b)-K. So when the loss
is high, the entropy coefficient is a small negative term, indicating
that entropy should be maximized. However, when the loss becomes very
low, the entropy term starts to grow above 0 and entropy is minimized.