Add adaptive length cost coef, mod symbol dropout
- I don't want to modify EGG's default wrappers too much, so I moved the
new wrapper in vd_reco/gs_wrappers.py
- Adaptive length cost coef = the coef in front of the length of message
penalty depends on the value of the loss, meaning that only on examples
where reconstruction is easy do we penalize the length
- I modified symbol dropout, but idk if it's useful: it replaces random
characters with an EOS token. I'd like to restore the other version
where it replaces random characters with another character, instead of
ending the message. That should be an option.

Here is an example of interesting HPs:

python -m egg.zoo.vd_reco.train_vl --vocab_size 16 --n_bits 8
--random_seed 0 --temperature 1 --batch_size 128 --lr 0.001
--sender_emb=30 --receiver_emb=30 --sender_hidden=120
--receiver_hidden=120 --sender_mlp --receiver_mlp --validation_freq 10
--train_test_ratio -1.0 --variable_bits --n_epochs 1000
--optimizer='smorms3' --sender_squash_output 0 --mode='gs'
--sender_cell='rnn' --receiver_cell='rnn' --max_len 4 --length_cost 0.03
--entropy_coef 0.00 --symbol_dropout 0.1 --ada_len_cost_thresh 0.05
--ada_H_cost_thresh 0.0