Unordered output + classical role + Hungarian
Instead of expecting the first role (if present) in the first position
in the output, it can be anywhere. The matching algorithm uses the cross
entropy loss on classical role as well as on all the outputs with EQUAL
weight (see loss_unordered).

Eventually, we can completely get rid of classical roles, instead
computing representations for each argument types based on clusterings.