Back to ordered loss with various fixes
- Predict whether the output object with role i is present or a
"padding" object. I fear this loss should have a higher coefficient than 1 b/c it's a bit
diluted in the cross-entropy of the features right now. And it is quite
important whether or not the object is there or not!
- The feature reconstruction cross entropy is masked for padding objects.
- The missing or hidden objects have a special, shared embedding (across
all positions) that indicate that they're missing (or hidden, for the
Listener). It makes sense that the listener does not know if some role
is hidden or not fulfilled. At least, the listener can now
distinguish "all-0s" objects from missing/hidden.