Fix proba EoS + add truncation option for concat
Bug: the sampling procedure is such that EoS position in the last
position has probability 1. Therefore, all sequences with length >
max_len have 0 probability. This was not the case in the function
to compute probability. The evaluation function was evaluating a
modified probability distribution that is useful to compute probability
of longer sequences, but that is NOT the training distribution.

The training distribution (`has_max_len=True`) is used for:

- concatenability metrics where concatenated messages are truncated to
  max len
- mutual information computation