Make predict n features work
Slight modification of the architecture: we add the feature markers only
for the 2nd transformer. We also use more heads & large learning rate.

To debug, I have implemented an auxiliary loss that predicts the number
of similar features to the target object (for each feature, how many are
identical), after the 1st Transformer. With a cross ent loss of around
0.55 we can already solve the task at 100% (K=1), 94% (K=2) at around
  iteration 40.