Fix causal bug + cleaner weight bias
Before bug fix, the `eval_proba_sender()` was incorrect for sender with
more than 1 layers. So this should yield better speaker log likelihood
(although I don't expect a lot of changes in terms of model ranking)

The weight bias is now implemented only on the weights and not directly
on query/key/value.